WIP: Search Bar UI Design (#39)
* Commiting long binary expression changed for
smells in project1 in src/privileges/topics.js

* Added a search bar into category.tpl for post
navigation. It currently opens a page not found
when clicked.

* Commented out visibility line because it was
causing errors.

* Removed visibility line comment code because no
longer working on backend.

* Adjusted search bar UI so that it doesn't redirect
to an page not found page when clicked. Removed the
old implementation of importing prexisting search bar.
Search bar has a button and can be typed into.

* Adding documentation for the search bar UI
implementation in category.tpl

* Reverting commit of complex binary expression fix
from project 1 in priviledges/topics.js as it will
be merged in through a different pull request from
a branch in the main repo.