Date Filtering Button UI (#43)
* Button for date filtering but currently leads to a
page not found page.

* Added Documentation for the button implementation
for topic-list-bar.tpl in nodebb-theme-harmony-2.1.15

* Debugged alignment issue with date filtering
button with RSS Feed icon in topic-list-bar.tpl.
Changing the bolding for text to use <span> instead of <strong>
to avoid layout issues.

* Fixed syntax error in topic-list-bar.tpl
where there was a closing > missing for data filtering button.
This resulted in the icon not being displayed properly.

* Added container toggled by the date filter Button
The script is inside of topic-list-bar.tpl at the bottom.
The styling is done inside of the template file with the "style" tag inside div"
The container contains 3 text inputs and a submit button.
Can be toggled on and off.

* Documentation + Bolding/Spacing Style Changes
I lowered the bolding on the Date Filtering Button to make it more
consistent with the other buttons in the topic list bar. I also just
added documentation for the Date Filtering Button on topic-list-bar.tpl